+++ src/features/transactions/components/app-call-transaction-info.tsx
@@ -8,9 +8,9 @@ import { DescriptionList } from '@/features/common/components/description-list'
 import { AccountLink } from '@/features/accounts/components/account-link'
 import { ApplicationLink } from '@/features/applications/components/application-link'
 import { applicationIdLabel } from '@/features/applications/components/labels'
-import { transactionSenderLabel } from './labels'
+import { abiMethodNameLabel, transactionSenderLabel } from './labels'
 import { AssetIdLink } from '@/features/assets/components/asset-link'
-import { DecodedAbiMethod } from '@/features/transactions/components/decoded-abi-method'
+import { DecodedAbiMethod } from '@/features/abi-methods/components/decoded-abi-method'
 import { isDefined } from '@/utils/is-defined'
 import { useAtomValue } from 'jotai'
 import algosdk from 'algosdk'
@@ -36,7 +36,6 @@ export const foreignAssetsTabLabel = 'Foreign Assets'
 export const globalStateDeltaTabLabel = 'Global State Delta'
 export const localStateDeltaTabLabel = 'Local State Delta'
 export const decodedAbiMethodTabLabel = 'Decoded ABI Method'
-export const abiMethodNameLabel = 'ABI Method Name'
 export const appCallTransactionDetailsLabel = 'App Call Transaction Details'
 export const onCompletionLabel = 'On Completion'
