+++ src/utils/base64-to-bytes.ts
@@ -1,0 +1,5 @@
+import { Buffer } from 'buffer'
+
+export const base64ToBytes = (arg: string) => {
+  return Uint8Array.from(Buffer.from(arg, 'base64'))
+}
